requirements.txt
pandas==1.3.4
numpy==1.21.2
requests==2.29.0
python-dotenv==0.20.0
google-cloud-bigquery==2.34.3
zlib==1.2.11
dataclasses-json==0.6.7
python-binance==1.0.16
web3==5.30.0
matplotlib==3.4.3
seaborn==0.11.2
jupyter==1.0.0
ipynb==0.5.1
